ai_agent/requirements.txt
# Web framework
flask==2.3.3
flask-cors==3.0.10

# HTTP requests
requests==2.32.0

# PDF processing
PyPDF2==3.0.1
pdf2image==1.16.3

# OCR
pytesseract==0.3.10

# Environment variables
python-dotenv==1.0.0

# Word processing
python-docx==1.2.0

# Excel support
pandas==2.1.1
numpy==1.26.1
openpyxl==3.1.2
xlrd==2.0.1

# LLM support
langchain==0.3.27
langchain-ollama==0.3.8

# Optional: image handling (Pillow is required by pdf2image)
Pillow==10.1.0
